requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-qdrant>=0.0.1
qdrant-client>=1.7.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.4.0
gitpython>=3.1.32
beautifulsoup4>=4.12.2
requests>=2.31.0
markdown>=3.4.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
